requirements.txt
numpy
pandas
matplotlib
osmnx
geopandas
shapely
pyosmium